pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lexiglow-backend"
version = "0.1.0"
description = "Backend for the LexiGlow application."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }                               # Or use the SPDX identifier: license = "MIT"
authors = [{ name = "Your Name", email = "your@email.com" }]

# Core dependencies for your application to run
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pymongo>=4.0",
  "motor>=3.0",
  "python-dotenv>=0.21",
  "SQLAlchemy>=2.0",
  "pydantic[email]>=2.0",
  "bcrypt>=4.0",
  "markdown>=3.10",
  "httpx>=0.25.0",
  "aiosqlite>=0.21.0",
  "python-ulid>=3.1.0",
]

[project.optional-dependencies]
# Dependencies for development, linting, and testing
dev = [
  "pytest>=7",
  "pytest-asyncio>=0.21",
  "ruff>=0.4.4",           # Replaces black and flake8
  "pre-commit>=3",
  "pdoc3>=0.10.0",         # For API documentation generation
  "mypy>=1.0",
  "types-Markdown>=0.1.0", # Type stubs for markdown
]

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
warn_return_any = true
warn_no_return = true
explicit_package_bases = true
namespace_packages = true
packages = ["app", "tests", "scripts"]

# Ignore missing type stubs for libraries that don't have them
[[tool.mypy.overrides]]
module = ["uvicorn.*", "motor.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
# Assume Python 3.12+ code. Adjust this to your project's minimum Python version.
target-version = "py312"
# indent-width = 4

[tool.ruff.lint]
# A good starting set of rules:
# E, F: Standard pycodestyle and Pyflakes
# I: isort (import sorting)
# UP: pyupgrade (modernize code)
# B: flake8-bugbear (finds likely bugs)
# C4: flake8-comprehensions (encourages more idiomatic comprehensions)
select = ["E", "F", "I", "UP", "B", "C4"]

# You can ignore specific rules if they are too noisy for your project.
# ignore = []

[tool.ruff.format]
# Use black-compatible formatting.
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "."
norecursedirs = "data"
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
